refactor(fuzz): split fuzz.rs command god-file into modules
